pyproject.toml
[project]
name = "agentku"
version = "0.1.0"
description = ""
authors = [
    {name = "deduu",email = "dedy.ariansyah1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.68.2,<2.0.0)",
    "playwright (>=1.51.0,<2.0.0)",
    "transformers (>=4.50.0,<5.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "torchaudio (>=2.6.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
